refactor(web): remove unnecessary `/universite/[slug]/*` pages
Since we don't have the ability to rate faculties, we don't need
to display them. That's why we removed the faculty and faculty list
pages.

Also moved index page of `/universite/[slug]`, which displays ratings
of the university, to `/universite/[slug]/degerlendirmeler` and add a
rewrite rule to `next.config.js` to redirect `/universite/[slug]` to
`/universite/[slug]/degerlendirmeler`. That way we can use same layout,
which has university information and base ratings, for both pages.