fix(web): create RateContext
RateContext allows us to store the rating data for university or
teacher in a single place. By doing this, we don't need to prop drill
the rating data (such as `ratingTo`).

We are also passing the rating data to the RateForm component by
creating another RateContext. This way, we can get the `ratingTo` data
from `/university/[slug]` or `/teacher/[slug]` pages.